pyproject.toml
[project]
name = "performant-rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-versioning>=0.10.0",
    "fastapi[standard]>=0.114.0",
    "fastenv>=0.5.0",
    "llama-index-embeddings-openai>=0.2.4",
    "llama-index-vector-stores-qdrant>=0.3.0",
    "llama-index>=0.11.7",
    "qdrant-client>=1.11.1",
    "ruff>=0.6.4",
]


[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "E712",  # Allow using if x == False, as it's not always equivalent to if x.
    "E501",  # Supress line-too-long warnings: trust black's judgement on this one.
    "UP017",  # Allow timezone.utc instead of datetime.UTC.
    "D100",
    "D101", # Allow no docstring
    "D103",
    "D104",
    "T201",
    "FBT001",
]

[tool.ruff]
line-length = 120


